move to flex package
